refactor: Create ViewModel for WordMatch Container
1. Move Business Logic in Word Match Container
